Update fetch_release.sh
Stop script on first error with set -e. If tar fails mv will fail because of that making it harder to understand what failed. On a side note guys, my install was failing because $npm_package_version was undefined ($1 to script). This cause curl to fetch a non .gz file causing tar to fail. Probably an issue with my env, not sure. Should I open an issue?